Move current_user logic to concern
